Move all App.js functionality into index.js
One of the requirements of js-digital-clock is a single JS file, so
here we go!
This also fixes a bug in which the module export of the App component
was busted, thus breaking the entire app. :(
